+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.1] - 2026-02-12
+
+- fix: only run radioplayer in non-prod or when requested
+- fix: only post dts and radioplayer queues for now playing events
+
 ## [2.3.0] - 2026-02-10
 
 - fix: ignore blocked services in common topic